added the rest of the onboarding pages
